requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
requests>=2.28.0
paho-mqtt>=1.6.0
kubernetes>=24.0.0
packaging>=21.0
tabulate>=0.9.0